+++ src/games/games.stories.tsx
@@ -131,6 +131,46 @@ export const Emscell: Story = {
       "col6",
       "col7",
       "col8"
-    ]
+    ],
+    onMove: (cardStacks, moveCardThunk) => {
+      const playStacks = Object.values(cardStacks).filter((stack) => !stack.name.startsWith("suit"));
+      const suitStacks = Object.values(cardStacks).filter((stack) => stack.name.startsWith("suit"));
+
+      let cardToMove: ICard | undefined, destinationStack: IStack | undefined;
+
+      playStacks.forEach((playStack) => {
+        if (cardToMove) {
+          // Short-circuit if we've already found a card to move.
+          return;
+        }
+        // Look at the last card of the stack.
+        const candidate = playStack.cards.length && playStack.cards[playStack.cards.length - 1];
+
+        if (!candidate) {
+          // If the stack's empty, move on.
+          return;
+        }
+
+        const availableSuitStack = suitStacks.find((suitStack) => {
+          const lastSuitCard = suitStack.cards.length && suitStack.cards[suitStack.cards.length - 1];
+
+          if (lastSuitCard && lastSuitCard.suit === candidate.suit && lastSuitCard.rank === candidate.rank - 1) {
+            return true;
+          } else if (!lastSuitCard && candidate.rank === 1) {
+            return true;
+          }
+        });
+
+        if (availableSuitStack) {
+          // We've got a match!
+          cardToMove = candidate;
+          destinationStack = availableSuitStack;
+        }
+      });
+
+      if (cardToMove && destinationStack) {
+        moveCardThunk({ card: cardToMove, toStack: destinationStack.name });
+      }
+    }
   }
 };