+++ src/playmat/Playmat.tsx
@@ -2,16 +2,17 @@ import { ActionCreator, StoreProvider } from "easy-peasy";
 import React, { FunctionComponent, ReactNode, useEffect } from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import { store, useStoreActions, useStoreState } from "../model/storeModel";
+import { OnMove, store, useStoreActions, useStoreState } from "../model/storeModel";
 
 interface Props {
   setup?: (cardStacks: CardStacks, moveCard: ActionCreator<MoveCardPayload>) => void;
-  isWin?: (cardStacks: CardStacks) => boolean;
+  isWin?: IsWin;
+  onMove?: OnMove;
   preferredMoveStacks?: string[];
   children?: ReactNode;
 }
 
-const PlaymatInner: FunctionComponent<Props> = ({ setup, isWin, preferredMoveStacks, children }) => {
+const PlaymatInner: FunctionComponent<Props> = ({ setup, isWin, onMove, preferredMoveStacks, children }) => {
   const cardStacks = useStoreState(
     (store) => store.cardStacks,
     (previous, next) => Object.keys(previous).length === Object.keys(next).length
@@ -20,11 +21,14 @@ const PlaymatInner: FunctionComponent<Props> = ({ setup, isWin, preferredMoveSta
   const storeIsWin = useStoreState((store) => store.isWin);
   const win = useStoreState((store) => store.win);
   const storePreferredMoveStacks = useStoreState((store) => store.preferredMoveStacks);
+  const storeOnMove = useStoreState((store) => store.onMove);
+
   const moveCard = useStoreActions((store) => store.moveCard);
   const setSetupHasRun = useStoreActions((store) => store.setSetupHasRun);
   const setIsWin = useStoreActions((store) => store.setIsWin);
   const undo = useStoreActions((store) => store.undo);
   const setPreferredMoveStacks = useStoreActions((store) => store.setPreferredMoveStacks);
+  const setOnMove = useStoreActions((store) => store.setOnMove);
 
   useEffect(() => {
     const listener = (event: KeyboardEvent) => {
@@ -48,6 +52,12 @@ const PlaymatInner: FunctionComponent<Props> = ({ setup, isWin, preferredMoveSta
       setPreferredMoveStacks(preferredMoveStacks);
     }
   }, [preferredMoveStacks]);
+
+  useEffect(() => {
+    if (!storeOnMove && onMove) {
+      setOnMove(onMove);
+    }
+  }, [onMove]);
 
   useEffect(() => {
     if (!setupHasRun && cardStacks && Object.keys(cardStacks).length) {